Rename artifact: "darcy-ui"
Darcy is a modular project, and this artifact is, and has been, the "UI"
module in that project. Where "web" adds web-specific DSL, UI defines
the API for universal UI elements. I expect other modules to come around
looking like "darcy-mobile" or "darcy-desktop" -- in other words, there
is a clear pattern, and this module is not following it.

I think it also helps communicate that this artifact is not the "darcy"
project -- it is merely a component of it.