Remove ProxyElementSelection and redundancy in InvocationHandlers
Since clarifying the behavior of `Context#find()` and `Locator#find`
the proxy element selection is no longer necessary. The caching behavior
has been moved implementation-side. The only reason we still need _some_
proxies are for element fields that do not have a context to be 'found'
in until `View#setContext(ElementContext)` is called.