Gut Chained/NestedElementContext proxy concept
It was bad and I feel bad.

You can't proxy contexts because the whole concept of contexts is that
they implement FindsByWhatevers, all of which are nontrivial to override
in the proxy to nest them underneath some other locator. Ultimately, the
alternative solution was to push the burden onto the implementations.
It's not too hard to implement, and is a lot less messy.