Remove final modifier from getContext
I'm conflicted on the value of final modifiers so I would not be
surprised if more started to get removed. In this case, it's nice for
specific Views to be able to override the return value of getContext
when you know the context for that view _must_ be a certain type, more
specific than ElementContext (such as WebContext).