Revert "Require parent element"
This reverts commit d09ee805561476199467cc7b5341cd4f30c0add1.

This make sense, but I don't like that it is not explicit. Additionally,
you can't take advantage of the analyzer if you don't want to require
the parent for whatever reason (although that might not make any sense).

More than that, I'm mainly concerned about the case where the only
required element is the parent and then there is nothing to indicate
that other than javadoc.