Only allow findables to be used with Locators (fixes #49)
This sets a contractual rule that helps make working with darcy more
consistent. Anything that is found from a locator must not be null, and
it must not throw an exception if it cannot find the thing. Whether or
not it is lazily found is up to the implementation. Essentially it means
that everything returned from a locator is optional, but instead of
wrapping in an Optional type and forcing folks to call `get()` in order
to access the type; they can already access the type, the type must
just adhere to the Findable interface contract and only throw up if
someone tries to use the type when it cannot be found (in other words,
when isPresent would return false).